feat(cache): Override and enhance make_and_validate_key method
This commit overrides the `make_and_validate_key` method. It extends the base class's key validation by adding a step to transform the prefix of the validated key into a directory-like structure. This aligns object keys with best practices for distributing write operations, which helps boost throughput and performance, especially for high-volume data