feat(cache): Override and enhance make_key method
This commit overrides the `make_key` method. It extends the base class's key method by adding a step to transform the prefix of the key into a directory-like structure. This aligns object keys with best practices for distributing write operations, which helps boost throughput and performance, especially for high-volume data